Refactor integration tests for MongoDB backend to use base tests
